Fix Issue #56 and add include guards
Issue #56 was caused by multiple redefinition of internal `lux/` files
when importing a header / cpp file combo that both contain
`lux/kit.hpp`. Specifically, it was caused by static variable
declarations in lux_io.cpp and constants.cpp and the operator <<
function in position.hpp.

To fix this, simply inline those functions and variable declarations.
The only downside to this is that the project must now be compiled with
C++ 17 in order for this to work.

In addition to this, add include guards to every single header file in
`lux/` for the sake of good practice.