Alternate fix for Issue #56
Here is an alternate fix for Issue #56 that implements the cpp file I
was talking about WITHOUT the use of C++17. Basically, a file named
`define.cpp` contains definitions for both the static variable
definitions of `lux_io.cpp` and the operator << function for
position.hpp. Because source files are only compiled once, this fixes
the issue of multile redefinitions. In order to compile, this file is
included in the compile command in both `compile.bat` and `compile.sh`,
though I haven't tested the .bat file on Windows.

A variable in `constants.hpp` however couldn't be fixed this way because
it wasn't static, meaning a copy of it was going to be in every file no
matter what. To fix that, I simply turned it static.